Refactor trafficrwc to wrappers
